+++ src/main/java/dev/celerbi/easyfarmersdelightcompat/registry/ModCreativeTabs.java
@@ -29,6 +29,7 @@ public final class ModCreativeTabs {
                         output.accept(ModBlocks.PADDY_FARMER_ITEM.get());
                         output.accept(ModBlocks.RICH_FARMER_ITEM.get());
                         output.accept(ModBlocks.RICH_PADDY_FARMER_ITEM.get());
+                        output.accept(ModBlocks.CUTTER_ITEM.get());
                     })
                     .build()
     );